+++ Ruleset/Configs/ServerConfig.cs
@@ -1,5 +1,4 @@
 ﻿using Newtonsoft.Json;
-using oomtm450PuckMod_Ruleset.SystemFunc;
 using System;
 using System.IO;
