+++ Ruleset/PuckRaycast.cs
@@ -1,5 +1,4 @@
-﻿using oomtm450PuckMod_Ruleset.SystemFunc;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
 namespace oomtm450PuckMod_Ruleset {
