+++ Ruleset/RefSignals.cs
@@ -1,5 +1,4 @@
-﻿using oomtm450PuckMod_Ruleset.SystemFunc;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
