+++ Ruleset/Ruleset.cs
@@ -1,8 +1,6 @@
 ﻿using HarmonyLib;
 using oomtm450PuckMod_Ruleset.Configs;
-using oomtm450PuckMod_Ruleset.SystemFunc;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
